requirements.txt
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
backcall==0.2.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
filelock==3.16.1
hyperlink==21.0.0
idna==3.10
importlib-metadata==8.5.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.12.3
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
jmespath==1.0.1
jupyter-client==8.6.3
jupyter-core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parsel==1.9.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.48
Protego==0.3.1
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pygments==2.18.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
stack-data==0.6.3
tldextract==5.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
twisted==24.7.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
wcwidth==0.2.13
zipp==3.20.2
zope.interface==7.0.3
